+++ server/flask_app/work/views.py
@@ -50,9 +50,10 @@ def delete_work(work_id, user_id, admin_override=False):
 		if work.user.id == user_id or admin_override:
 			db.session.delete(work)
 			db.session.commit()
-			doc = WorkSearch.get(id=work_id)
-			if doc is not None:
-				doc.delete()
+			if app.config.get('USE_ES'):
+				doc = WorkSearch.get(id=work_id)
+				if doc is not None:
+					doc.delete()
 			return work_id
 
 def update_work(json):
